Feature epayco (#90)
* pass the information client

* fix and pass all the data and get the response

* fix the path for response

* deleted console dir and put the env for url payment
