+++ src/pages/ReservationConfirm.js
@@ -1,43 +1,16 @@
 import { ReservationConfirmPayment } from '../components/Reservation/ReservationConfirmPayment'
-import { useSelector } from 'react-redux'
-import axios from 'axios'
+import { useSelector, useDispatch } from 'react-redux'
 import { useEffect } from 'react'
+import { getClient } from '../store/actions/Client.actions'
 
 const handler = window.ePayco.checkout.configure({
   key: process.env.REACT_APP_EPAYCO_PUBLIC_KEY,
   test: true
 })
 
-function queryString(query) {
-  const res = {}
-  query
-    .replace('?', '')
-    .split('&')
-    .forEach(q => {
-      const [key, value] = q.split('=')
-      res[key] = value
-    })
-  return res
-}
-
-function Response({ location }) {
-  useEffect(() => {
-    const { ref_payco } = queryString(location.search)
-
-    axios({
-      method: 'GET',
-      baseURL: 'https://api.secure.payco.co',
-      url: `/validation/v1/reference/${ref_payco}`
-    })
-      .then(({ data }) => {
-        console.log(data)
-      })
-  }, [location])
-
-  return <p>Repuesta de la transacción</p>
-}
-
 function ReservationConfirm() {
+
+  const dispatch = useDispatch()
 
   const today = new Date()
   const month = today.toLocaleString('es-CO', { month: 'long'}) 
@@ -50,6 +23,25 @@ function ReservationConfirm() {
     }}) => {
       return { ...state} 
     })
+
+  useEffect(() => {
+    dispatch(getClient())
+  }, [])
+
+  const {data} = useSelector(
+    ({ clientReducer: {
+      data
+    }}) => {
+      return {
+        data
+      }
+  })
+
+  const clientData = data
+
+  const invoice = localStorage.getItem('reservation');
+
+  const URL = process.env.REACT_APP_BASE_URL || 'http://localhost:3000'
   
   const handleClick = () => {
     handler.open(
@@ -57,7 +49,7 @@ function ReservationConfirm() {
         //Parametros compra (obligatorio)
         name: `${nameRestaurantReservation}`,
         description: "Pago Reserva alamesa",
-        invoice: "#RESERVA ",
+        invoice: `${invoice}`,
         currency: "cop",
         amount: `${deposit}`,
         tax_base: "0",
@@ -71,18 +63,18 @@ function ReservationConfirm() {
 
         //Atributos opcionales
         extra1: "descripción reserva",
-        response: `${process.env.REACT_APP_BASE_URL}/response`,
+        response: `${URL}/response`,
 
         //Atributos cliente
-        name_billing: "Nombre del cliente",
-        address_billing: "dirección cliente",
-        type_doc_billing: "cc",
-        mobilephone_billing: "1098971546",
-        number_doc_billing: "3104567891",
+        email_billing: `${clientData.email}`,
+        name_billing: `${clientData.name} ${clientData.lastName}`,
+        address_billing: `${clientData.address}`,
+        type_doc_billing: "CC",
+        mobilephone_billing: `${clientData.phone}`,
+        number_doc_billing: `${clientData.identification}`,
 
        //atributo deshabilitación metodo de pago
         methodsDisable: ["SP","CASH"]
-
       }
     )
   }
